% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_tracking_functions.R
\name{allele_track_back}
\alias{allele_track_back}
\title{Function to track gene-dropped alleles back through ancestors}
\usage{
allele_track_back(id, loci, gene_drop_out)
}
\arguments{
\item{id}{A vector containing numerical or character values identifying individuals
in the pedigree}

\item{loci}{A vector containing numerical values relating to the loci of interest}

\item{gene_drop_out}{A gene_drop_object from gene-dropping}
}
\description{
A function that tracks provided gene-dropped individuals' alleles at requested loci back
to the ancestor in which they originated.  A list of allele_track_objects is returned.
}
\examples{

}
\keyword{allele}
\keyword{ancestor}
\keyword{haplotype}
\keyword{track}
