% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_tracking_functions.R
\name{allele_track_back_sing}
\alias{allele_track_back_sing}
\title{Function to track a gene-dropped allele back through ancestors}
\usage{
allele_track_back_sing(id, loci, gene_drop_out)
}
\arguments{
\item{id}{A single character or numeric value identifying an individual in the pedigree}

\item{loci}{A single numeric value relating to a locus of interest}

\item{gene_drop_out}{A gene_drop_object from gene-dropping}
}
\description{
A function that tracks a gene-dropped individuals alleles at a locus back to the ancestor in
which they originated.
}
\examples{

}
\keyword{allele}
\keyword{ancestor}
\keyword{haplotype}
\keyword{track}
