% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_tracking_functions.R
\name{allele_track_forw}
\alias{allele_track_forw}
\title{Function to track gene-dropped alleles forward through descendants}
\usage{
allele_track_forw(id, loci, gene_drop_out)
}
\arguments{
\item{id}{A vector containing numerical or character values identifying individuals
in the pedigree}

\item{loci}{A vector containing numerical values relating to the loci of interest}

\item{gene_drop_out}{A gene_drop_object from gene-dropping}
}
\description{
A function that tracks provided gene-dropped individuals' alleles at requested loci forward
through their descendants.  A list of allele_track_objects is returned.
}
\examples{

}
\keyword{allele}
\keyword{descendant}
\keyword{haplotype}
\keyword{track}
