% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_tracking_functions.R
\name{allele_track_forw_sing}
\alias{allele_track_forw_sing}
\title{Function to track a gene-dropped allele forward through descendants}
\usage{
allele_track_forw_sing(id, loci, gene_drop_out)
}
\arguments{
\item{id}{A single character or numeric value identifying an individual in the pedigree}

\item{loci}{A single numeric value relating to a locus of interest}

\item{gene_drop_out}{A gene_drop_object from gene-dropping}
}
\description{
A function that tracks a gene-dropped individuals alleles at a locus forward through all the descendants
which received them.
}
\examples{

}
\keyword{allele}
\keyword{descendant}
\keyword{haplotype}
\keyword{track}
