% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_fix_fuctions.R
\name{calc_ped_depth}
\alias{calc_ped_depth}
\title{Function to calculate pedigree depth}
\usage{
calc_ped_depth(pedigree)
}
\arguments{
\item{pedigree}{A dataframe or matrix. A pedigree with ID, Sire and Dam columns}
}
\description{
This function calculates pedigree depth each individual in a pedigree by finding the
number to descendant generations
}
\examples{

}
\keyword{complete}
\keyword{fix}
\keyword{pedigree}
