% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_dropping_functions.R
\name{calc_recom_freq}
\alias{calc_recom_freq}
\title{Function to calculate recombination frequency}
\usage{
calc_recom_freq(map_dist)
}
\arguments{
\item{map_dist}{A kosambi map distance in cM}
}
\description{
A function to calculate recombination frequency based on kosambi map distances
}
\examples{

}
\keyword{frequency}
\keyword{kosambi}
\keyword{recombination}
