% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_fix_fuctions.R
\name{cohort_numeric}
\alias{cohort_numeric}
\title{Function to convert cohorts to numeric values}
\usage{
cohort_numeric(cohorts)
}
\arguments{
\item{cohorts}{A vector or matrix holding cohort information}
}
\description{
This function trys to return a vector of numerical values
fora cohort and produces an error if this isn't possible
}
\examples{

}
\keyword{check}
\keyword{cohort}
\keyword{numeric}
