% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_fix_fuctions.R
\name{est_cohort}
\alias{est_cohort}
\title{Function to estimate missing cohorts}
\usage{
est_cohort(peduse, tt_rep = 2)
}
\arguments{
\item{peduse}{A dataframe or matrix. A pedigree with ID, Sire and Dam columns}

\item{tt_rep}{integer. Time to reproduction.  A value of the length of time between
birth and reproduction. Defaults to 2}
}
\description{
A function that estimates cohorts for individuals without a known cohort.
It uses only pedigree and cohort information and should be secondary to more
accurate predictions.  If offspring information is present this is used, based on a provided
to reproduction.  When offspring data is absent a mean of sibling cohorts is used.
The function loops using the new estimates until no further cohorts can be
estimated.
}
\keyword{cohort}
\keyword{estimate}
