% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_functions.R
\name{find_least_common}
\alias{find_least_common}
\title{Function to find least common value in a vector}
\usage{
find_least_common(values)
}
\arguments{
\item{values}{vector. A vector of values}
}
\description{
A function that finds which values in a vector which are least common
}
\examples{

}
\keyword{common}
\keyword{least}
\keyword{values}
