% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extra_functions.R
\name{find_most_common}
\alias{find_most_common}
\title{Function to find most common value in a vector}
\usage{
find_most_common(values)
}
\arguments{
\item{values}{vector. A vector of values}
}
\description{
A function that finds vhich values in a vector which are most common
}
\examples{

}
\keyword{complete}
\keyword{fix}
\keyword{pedigree}
