% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_fix_fuctions.R
\name{fix_pedigree}
\alias{fix_pedigree}
\title{Function to ensure pedigree is correctly structured for gene dropping}
\usage{
fix_pedigree(peduse, sex = TRUE, cohort = TRUE)
}
\arguments{
\item{peduse}{A dataframe or matrix. A pedigree with ID, Sire and Dam columns.}

\item{sex}{logical or vector. TRUE if sex column is included in pedigree, otherwise
a vector containing the sex of each individual. Defaults to TRUE}

\item{cohort}{logical or vector. TRUE if cohort column is included in pedigree, otherwise
a vector containing the cohort of each individual. Defaults to TRUE}
}
\description{
This function checks the pedigree for potentential problems in structure
it reorganises the individuals if neccessary and renames columns to ensure the
pedigree is in the correct format for gene-dropping.  Sex and Cohort can be
included as columns in the pedigree or provided as seperate vectors.
It adds in any individuals that are missing and checks sex coding is correct
and consistent. It returns the pedigree as a matrix
}
\keyword{complete}
\keyword{fix}
\keyword{pedigree}
\keyword{structure}
