% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_dropping_functions.R
\docType{class}
\name{gene_drop_object-class}
\alias{gene_drop_object-class}
\alias{gene_drop_object}
\alias{id_info}
\alias{id_ref}
\alias{genotype_mat}
\title{An S4 class to store gene-dropping output}
\usage{
id_info(gene_drop_object, ids)

id_ref(gene_drop_object, ids)

genotype_mat(gene_drop_object, ids = "ALL", loci = "ALL")
}
\arguments{
\item{ids}{A scalar or vector of IDs to include. Defaults to ALL}

\item{loci}{A scalar or vector of loci to include. Defaults to ALL}
}
\description{
An S4 class to store gene-dropping output
}
\section{Functions}{
\itemize{
\item \code{id_info}: A method to extract the indivdual information from the pedigree

\item \code{id_ref}: A method to get and individuals row reference from the pedigree

\item \code{genotype_mat}: A method to return a genotype matrix
}}

\section{Slots}{

\describe{
\item{\code{genotype_matrix}}{A list of vectors each storing a haplotype}

\item{\code{haplotype_info}}{A list containing haplotype information for tracing alleles}

\item{\code{pedgree}}{The pedigree that was passed to the gene-dropping function}
}}

