% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_dropping_functions.R
\name{genedrop}
\alias{genedrop}
\title{Function to perform gene-dropping}
\usage{
genedrop(
  pedigree,
  map_dist,
  chr_loci_num,
  found_hap,
  to_raw = TRUE,
  sample_hap = TRUE,
  recom_freq = "kosambi"
)
}
\arguments{
\item{pedigree}{A dataframe or matrix. A pedigree containing 'ID', 'Sire' and 'Dam' columns}

\item{map_dist}{A vector of map distances in cM. The vector should be the same length as the total number of loci in
chr_loci_num.  Each value should be the difference in cM between the two markers, such that the first value in the vector
is the difference between marker 1 and marker 2.  The final value in the vector is therefore redundant but should be included
(e.g. as 0). Alternatively can be set to  NULL if recombination frequencies are provided using recom_freq.}

\item{chr_loci_num}{A vector containing the number of loci on each chromosome}

\item{found_hap}{A vector of founder haplotypes, if sample_hap=FALSE the pair of haplotypes for each individual should
in the same order as in the provided pedigree}

\item{to_raw}{logical TRUE or FALSE.  If TRUE the genotypes is stored are raw vectors.  This allows quicker gene-dropping
and reduces the size of the output.  It must be FALSE if alleles do not all have numeric values between 0 and 255}

\item{sample_hap}{logical TRUE or FALSE.  If TRUE founder haplotypes are assigned to founders at random.  If FALSE
pairs of haplotypes are assigned to founders in the order they appear in the pedigree.  When FALSE the number of haplotypes in found_hap
must be twice the numeber of founders present in the pedigree}

\item{recom_freq}{'kosambi' is default, this setting calculates recombination frequencies based using the map distances
provided in map_dist.  Alternatively a vector of the recombination frequencies between each marker can be provided.
The vector should be the same length as the total number of loci in chr_loci_num  The first value in the vector should be
the recombination frequency between markers 1 and 2. The final value in the vector is therefore redundant but should be included
(e.g. as 0.  Values between chromosomes are best set to 0.5 although this happens automaticaly during gene-dropping.}
}
\description{
This function carries out gene-dropping down a provided pedigree
}
\keyword{allele}
\keyword{gene-dropping}
\keyword{recombination}
\keyword{segregation}
\keyword{tracking}
