% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\name{GeneDropEx_hap}
\alias{GeneDropEx_hap}
\title{Dataset GeneDropEx_hap}
\description{
This matrix contains a simulated haplotypes for use in gene-dropping
}
\details{
This is simulated haplotype data loosely based on the population history
of the soay sheep on St Kilda.

It contains 650 haplotypes each with 30000 alleles

Please feel free to play around with this data and let me know how realistic, or otherwise,
you think it is.
}
