% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\name{GeneDropEx_map}
\alias{GeneDropEx_map}
\title{Dataset GeneDropEx_map}
\description{
This dataframe contains a simulated mapping information for use in gene-dropping
}
\details{
This is simulated map data that describes the physical and genetic location of the loci
in the GeneDropEx_hap file. The distances are loosely based that observed in the
St Kilda soay sheep population.

Please feel free to play around with this data and let me know how realistic, or otherwise,
you think it is.
}
