% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset_documentation.R
\name{GeneDropEx_ped}
\alias{GeneDropEx_ped}
\title{Dataset GeneDropEx_ped}
\description{
This dataframe contains a simulated pedigree for use in gene-dropping
}
\details{
This is simulated pedigree data for a population that
experiences density dependant selection and has modest migration.

It is loosely based on the dynamics seen in the data collected by the
Soay Sheep Project.

There is a small amount of missing data.

It spans 35 years and includes 10000 individuals.  Migrants do not have a known cohort, but
the time that they arrived in the population can be found in the 'Arrived' column.

Please feel free to play around with this data and let me know how realistic, or otherwise,
you think it is.
}
