% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allele_tracking_functions.R
\name{offspring_with_allele}
\alias{offspring_with_allele}
\title{Function to find offspring that have a parents allele}
\usage{
offspring_with_allele(id, loci, hap, gene_drop_out)
}
\arguments{
\item{id}{A vector containing a single numerical or character value identifying the individual
of interest}

\item{loci}{A vector containing single numerical value relating to the loci of interest}

\item{hap}{The haplotype on which the allele of interest is located (Sire : 1, Dam : 2)}

\item{gene_drop_out}{A gene_drop_object from gene-dropping}
}
\description{
A function that finds all an individuals' offspring that received a given alleles.
Returns a matrix with containing offspring ID, Sex and which haplotype the allele
is on.
}
\keyword{allele}
\keyword{haplotype}
\keyword{offspring}
\keyword{track}
