% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_fix_fuctions.R
\name{ped_col_present}
\alias{ped_col_present}
\title{Function to check  pedigree columns}
\usage{
ped_col_present(peduse, check_column = c())
}
\arguments{
\item{peduse}{A dataframe or matrix. A Pedigree}
}
\description{
This function checks if ID, Sire and Dam Columns,
and any other required columns are present in the pedigree
}
\keyword{check}
\keyword{pedigree}
\keyword{structure}
