% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_allele_track.R
\name{plot_allele_desc}
\alias{plot_allele_desc}
\title{Function to plot inheritance of alleles}
\usage{
plot_allele_desc(
  gene_drop_object,
  id,
  loci,
  col_sire = "#f1a340",
  col_dam = "#998ec3",
  line_trans = 0.6,
  point_trans = 1,
  focal_cex = 1.5,
  point_cex = 0.6,
  background = TRUE,
  cohort_labels = TRUE,
  id_labels = FALSE
)
}
\arguments{
\item{gene_drop_object}{A gene_drop_object from gene-dropping}

\item{id}{A numerical or character scalar of the focal ID to plot.}

\item{loci}{A numeric scalar of the locus to plot.}

\item{col_sire}{A character scalar determining the solid colour used for sires.  Should be a colour name, integer, or hexadecimal string of the form "#rrggbb".
The default is "#f1a340".}

\item{col_dam}{A character scalar determining the solid colour used for dams.  Should be a colour name, integer, or hexadecimal string of the form "#rrggbb".
The default is "#998ec3".}

\item{line_trans}{A numeric scalar between 0 and 1 determining the transparency of the plotted lines. The default is 0.6,
setting to 0 hides the points.}

\item{point_trans}{A numeric scalar between 0 and 1 determining the transparency of the plotted points. The default is 1,
setting to 0 hides the points.}

\item{focal_cex}{A numeric scalar determining the size of the point for the focal individual. The default is 1.5.}

\item{point_cex}{A numeric scalar determining the size of the points used for the descendants. The default is 0.6, setting
this value to 0 allows only the point of the focal individual to be plotted.}

\item{background}{logical TRUE or FALSE. When TRUE the complete pedigree is plotted in light grey.  Default is TRUE but disabling will
make plotting much faster.}

\item{cohort_labels}{logical TRUE or FALSE. When TRUE cohort labels are printed down the left-hand side. Default is TRUE.}

\item{id_labels}{logical TRUE or FALSE. When TRUE id labels are printed above each point. Default is FASLE.}
}
\description{
Function to plot inheritance of both alleles present at a locus in an individual from a gene_drop_object
}
\examples{

}
\keyword{allele}
\keyword{descendant}
\keyword{plot}
\keyword{track}
