% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_dropping_functions.R
\name{reduce_hap_code}
\alias{reduce_hap_code}
\title{Function to condense haplotype source information}
\usage{
reduce_hap_code(sire_hap_info, dam_hap_info)
}
\arguments{
\item{sire_hap_info}{A vector of 1's and 2's determining sire haplotype source information}

\item{dam_hap_info}{A vector of 1's and 2's determining dam haplotype source information}
}
\description{
A function to condense haplotype source information.
A string of 1's and 2's determinining which parental haplotype
each allele of the desecendants  haplotype came from are condensed.
i.e. (111112222 becomes 1(5)2(4) but in matrix format)
A list of two matrices is returned, one relating to the sire and one to the dam.
}
\keyword{condense}
\keyword{haplotype}
\keyword{information}
\keyword{rle}
