% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_dropping_functions.R
\docType{class}
\name{gene_drop_object-class}
\alias{gene_drop_object-class}
\alias{gene_drop_object}
\alias{get_pedigree}
\alias{get_haplotype_info}
\alias{get_genotype_matrix}
\alias{id_info}
\alias{id_ref}
\alias{extract_genotype_mat}
\title{An S4 class to store gene-dropping output}
\usage{
get_pedigree(gene_drop_object)

get_haplotype_info(gene_drop_object)

get_genotype_matrix(gene_drop_object)

id_info(gene_drop_object, ids)

id_ref(gene_drop_object, ids)

extract_genotype_mat(gene_drop_object, ids = "ALL", loci = "ALL")
}
\arguments{
\item{gene_drop_object}{A gene_drop_object}

\item{ids}{A scalar or vector of IDs to include. Defaults to ALL}

\item{loci}{A scalar or vector of loci to include. Defaults to ALL}
}
\description{
An S4 class to store gene-dropping output
}
\section{Functions}{
\itemize{
\item \code{get_pedigree}: A method to access the pedigree of the gene-drop object

\item \code{get_haplotype_info}: A method to access the haplotype source information of the gene-drop object

\item \code{get_genotype_matrix}: A method to access the genotype matrix of the gene-drop object

\item \code{id_info}: A method to extract the indivdual information from the pedigree

\item \code{id_ref}: A method to get and individuals row reference from the pedigree

\item \code{extract_genotype_mat}: A method to return a genotype matrix
}}

\section{Slots}{

\describe{
\item{\code{genotype_matrix}}{A list of vectors each storing a haplotype}

\item{\code{haplotype_info}}{A list containing haplotype information for tracing alleles}

\item{\code{pedgree}}{The pedigree that was passed to the gene-dropping function}
}}

\examples{

}
