% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pedigree_fix_fuctions.R
\name{test_order}
\alias{test_order}
\title{Function to check pedigree order}
\usage{
test_order(pedigree)
}
\arguments{
\item{pedigree}{A dataframe or matrix. A pedigree containing 'ID', 'Sire' and 'Dam' columns}
}
\description{
This function checks if sires and dams are appearing after offspring
and returns TRUE or FALSE.
}
\examples{

}
\keyword{order}
\keyword{pedigree}
\keyword{sorting}
